backend/requirements.txt
Flask==2.0.2
Flask-CORS==3.0.10
torch==1.10.0
torchvision==0.11.1
Pillow==8.3.2
